refactor(unity-exec): split plan token coordinator responsibilities
Why:
- PlanTokenCoordinator held codec, key store, mode resolution, and fingerprint logic in one class, reducing cohesion and testability.

What Changed:
- Extracted compact token codec, key store, mode resolver, state fingerprint calculator, and token data models into dedicated files.
- Replaced the previous monolithic contracts file with per-type files and renamed the environment implementation to DefaultPlanTokenEnvironment.
- Kept coordinator as orchestration only, delegating issue/validate internals to focused collaborators.

Impact:
- No contract behavior change intended; internal structure is now separated by responsibility for safer maintenance and targeted tests.
