+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/DefaultPlanTokenEnvironment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fdfa044dee2f243b9a6632ae2531f7c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
