+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/IPlanTokenCoordinator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: caa132cac89704d03a2a05ada5ba2197
+guid: d78683a7227d64371a6dc68df52cb6ac
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
