+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/IPlanTokenEnvironment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1552edd70a934177a789e38656aa418
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
