+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenCompactCodec.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77fd3b0cde3b644139847a79204fe9c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
