+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenEnvironmentSnapshot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49fd76c62732849da9f4b6399980e456
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
