+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenMode.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace MackySoft.Ucli.Unity.Execution.PlanToken
+{
+    /// <summary> Defines resolved plan-token modes used by call validation. </summary>
+    internal enum PlanTokenMode
+    {
+        Optional = 0,
+        Required = 1,
+    }
+}
