+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenMode.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c136b55082b9249de9f4834fc3120b41
+guid: 6f0bc58eb958046c4aa874e552b22802
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
