+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenStateFingerprintCalculator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f147a0a927642421b922ded0759b895e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
