+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/PlanTokenTouchedDigestEntry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85c4b3514a1cd4e1f8304beb41fd67b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
